amplify/functions/use_aws/requirements.txt
strands-agents-tools==0.1.3